speculative -strict compile error fixes
#ROBOMERGE-SOURCE: CL 12216186
#ROBOMERGE-BOT: (v662-12191386)

#rb none
#ushell-cherrypick of 12216222 by david.ratti

[CL 12245696 by paul chipchase in Dev-Core branch]

+++ Engine/Plugins/Runtime/NetworkPredictionInsights/Source/NetworkPredictionInsights/Public/INetworkPredictionProvider.h
@@ -509,7 +509,7 @@ struct FSimulationData
 			: Store{Allocator, Allocator, Allocator}
 		{ }
 
-		FUserStateStore Store[ENP_UserState::MAX];
+		FUserStateStore Store[(int32)ENP_UserState::MAX];
 	};
 
 	struct FRestrictedUserStateView
